Fix a bug where folder would not expand if there was a space
We were expanding the folder divs using an id, which used to be constructed from the folder name, however if the name had a space in then the id would not work properly.
Switched to using the folder key instead. To do this, the route had to be changed to pass the whole folder model, not just the name.